Refactor: 로그인한 사용자가 작성한 글 조회 study api -> mypage api로 수정
